fix: ROUND_END시 기존의 stream을 destroy 하도록 변경
